[android] powermanagement: move implementation to platform folder
