+++ xbmc/platform/android/powermanagement/AndroidPowerSyscall.cpp
