+++ xbmc/platform/android/powermanagement/AndroidPowerSyscall.h
