+++ xbmc/platform/android/powermanagement/CMakeLists.txt
@@ -2,4 +2,4 @@ set(SOURCES AndroidPowerSyscall.cpp)
 
 set(HEADERS AndroidPowerSyscall.h)
 
-core_add_library(powermanagement_android)
+core_add_library(platform_android_powermanagement)
