+++ xbmc/windowing/android/WinSystemAndroid.cpp
@@ -40,7 +40,7 @@
 #include "cores/VideoPlayer/DVDCodecs/Audio/DVDAudioCodecAndroidMediaCodec.h"
 #include "cores/VideoPlayer/VideoRenderers/HwDecRender/RendererMediaCodec.h"
 #include "cores/VideoPlayer/VideoRenderers/HwDecRender/RendererMediaCodecSurface.h"
-#include "powermanagement/android/AndroidPowerSyscall.h"
+#include "platform/android/powermanagement/AndroidPowerSyscall.h"
 #include "addons/interfaces/platform/android/System.h"
 
 #include <EGL/egl.h>
